wip Fix some circular import issues
