+++ src/vercel/blob/ops.py
@@ -2,14 +2,18 @@ from __future__ import annotations
 
 from collections.abc import AsyncIterator, Awaitable, Callable, Coroutine, Iterable, Iterator
 from os import PathLike
-from typing import Any, TypeVar, cast
+from typing import Any, TypeVar
 
-from vercel._internal.iter_coroutine import iter_coroutine
+from vercel._internal.blob import (
+    ensure_token,
+    validate_access,
+)
 from vercel._internal.blob.core import (
     _AsyncBlobOpsClient,
     _SyncBlobOpsClient,
     normalize_delete_urls,
 )
+from vercel._internal.iter_coroutine import iter_coroutine
 from vercel.blob.types import (
     Access,
     CreateFolderResult as CreateFolderResultType,
@@ -19,10 +23,6 @@ from vercel.blob.types import (
     ListBlobResult as ListBlobResultType,
     PutBlobResult as PutBlobResultType,
     UploadProgressEvent,
-)
-from vercel._internal.blob import (
-    ensure_token,
-    validate_access,
 )
 
 _T = TypeVar("_T")
