+++ src/vercel/cache/cache_in_memory.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from collections.abc import Sequence
 
 from vercel._internal.telemetry.tracker import track
+
 from .types import AsyncCache, Cache
