+++ tests/integration/test_blob_sync_async.py
@@ -9,6 +9,7 @@ import httpx
 import pytest
 import respx
 
+from vercel._internal.blob.core import decode_blob_response_json
 from vercel.blob import (
     AsyncBlobClient,
     BlobClient,
@@ -34,7 +35,6 @@ from vercel.blob import (
     upload_file,
     upload_file_async,
 )
-from vercel._internal.blob.core import decode_blob_response_json
 from vercel.blob.ops import get, get_async
 
 # Base URL for Vercel Blob API
